refactor: implemented small changes to improve project structure
