+++ gradle/libs.versions.toml
@@ -77,9 +77,9 @@ ui-graphics = { group = "androidx.compose.ui", name = "ui-graphics", version.ref
 ads-mobile-sdk = { group = "com.google.android.libraries.ads.mobile.sdk", name = "ads-mobile-sdk", version.ref = "adsMobileSdk" }
 
 [plugins]
-android-application    = { id = "com.android.application",        version.ref = "agp" }
-kotlin-android        = { id = "org.jetbrains.kotlin.android",   version.ref = "kotlin" }
-kotlin-compose        = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
-dagger-hilt-android   = { id = "com.google.dagger.hilt.android", version.ref = "hiltAndroid" }
+android-application = { id = "com.android.application", version.ref = "agp" }
+kotlin-android = { id = "org.jetbrains.kotlin.android", version.ref = "kotlin" }
+kotlin-compose = { id = "org.jetbrains.kotlin.plugin.compose", version.ref = "kotlin" }
+dagger-hilt-android = { id = "com.google.dagger.hilt.android", version.ref = "hiltAndroid" }
